🧰👤: decouple flap component from sidebars
